refactor: rename error_helper exception_helper
